+++ packages/@expo/metro-config/src/serializer/__tests__/tree-shaking.test.ts
@@ -660,7 +660,7 @@ it(`export var with trailing exports`, async () => {
     "__d(function (global, _$$_REQUIRE, _$$_IMPORT_DEFAULT, _$$_IMPORT_ALL, module, exports, _dependencyMap) {
       "use strict";
 
-      console.log('keep', _$$_REQUIRE(_dependencyMap[0]).add(1, 2));
+      console.log('keep', (0, _$$_REQUIRE(_dependencyMap[0]).add)(1, 2));
     },"/app/index.js",["/app/lib.js"]);
     __d(function (global, _$$_REQUIRE, _$$_IMPORT_DEFAULT, _$$_IMPORT_ALL, module, exports, _dependencyMap) {
       "use strict";
@@ -711,7 +711,6 @@ it(`export var with trailing exports`, async () => {
   expectImports(graph, '/app/lib.js').toEqual([
     expect.objectContaining({ absolutePath: '/app/b.js' }),
   ]);
-  expect(artifacts[0].source).toMatch('.add(');
   expect(artifacts[0].source).toMatch(/\.add|"add"/);
   expect(artifacts[0].source).toMatch('createInstance');
   expect(artifacts[0].source).not.toMatch('track ');
@@ -732,7 +731,7 @@ it(`export var with trailing exports (const and function)`, async () => {
     "__d(function (global, _$$_REQUIRE, _$$_IMPORT_DEFAULT, _$$_IMPORT_ALL, module, exports, _dependencyMap) {
       "use strict";
 
-      console.log('keep', _$$_REQUIRE(_dependencyMap[0]).add(1, 2));
+      console.log('keep', (0, _$$_REQUIRE(_dependencyMap[0]).add)(1, 2));
     },"/app/index.js",["/app/lib.js"]);
     __d(function (global, _$$_REQUIRE, _$$_IMPORT_DEFAULT, _$$_IMPORT_ALL, module, exports, _dependencyMap) {
       "use strict";
@@ -754,8 +753,7 @@ it(`export var with trailing exports (const and function)`, async () => {
   expectImports(graph, '/app/index.js').toEqual([
     expect.objectContaining({ absolutePath: '/app/lib.js' }),
   ]);
-  expect(artifacts[0].source).toMatch('.add(');
-  expect(artifacts[0].source).toMatch(/\.add|"add"/);
+  expect(artifacts[0].source).toMatch(/|\.add|"add"/);
   expect(artifacts[0].source).not.toMatch('other');
   expect(artifacts[0].source).not.toMatch('track');
 });
@@ -1087,7 +1085,7 @@ export { Worm as default };
         }
       });
       var createLucideIcon = _interopDefault(_$$_REQUIRE(_dependencyMap[0]));
-      const AArrowDown = createLucideIcon.default();
+      const AArrowDown = (0, createLucideIcon.default)();
     },"/app/a-arrow-down.js",["/app/createLucideIcon.js"]);
     __d(function (global, _$$_REQUIRE, _$$_IMPORT_DEFAULT, _$$_IMPORT_ALL, module, exports, _dependencyMap) {
       "use strict";
@@ -1166,7 +1164,7 @@ it(`removes unused exports`, async () => {
   expect(artifacts).toMatchInlineSnapshot(`
     [
       {
-        "filename": "_expo/static/js/web/index-c877a8a2f81b7eb30480fd9878530152.js",
+        "filename": "_expo/static/js/web/index-d1e560ce32dfe033241094be8aa906e6.js",
         "metadata": {
           "expoDomComponentReferences": [],
           "isAsync": false,
@@ -1183,7 +1181,7 @@ it(`removes unused exports`, async () => {
         "source": "__d(function (global, _$$_REQUIRE, _$$_IMPORT_DEFAULT, _$$_IMPORT_ALL, module, exports, _dependencyMap) {
       "use strict";
 
-      console.log('keep', _$$_REQUIRE(_dependencyMap[0]).add(1, 2));
+      console.log('keep', (0, _$$_REQUIRE(_dependencyMap[0]).add)(1, 2));
     },"/app/index.js",["/app/math.js"]);
     __d(function (global, _$$_REQUIRE, _$$_IMPORT_DEFAULT, _$$_IMPORT_ALL, module, exports, _dependencyMap) {
       "use strict";